Squashed commit of the following:
commit d8bf4b598770ba43914507c5d19326b87388e6d1
Author: electric-el <electric-el@electric-eloquence.io>
Date:   Thu Jul 14 17:32:44 2022 -0400

    better sanitation

commit fcff05e6d3eb2c493c178991951bd0ae8df22684
Author: electric-el <electric-el@electric-eloquence.io>
Date:   Thu Jul 14 15:36:32 2022 -0400

    better debugging

commit 89ffee0160d3cf62e6ad6fb8714dd7771d817593
Author: electric-el <electric-el@electric-eloquence.io>
Date:   Thu Jul 14 15:19:03 2022 -0400

    updated base profile

commit 311818eab140f7c3d34904bf2213672b299a20b7
Author: electric-el <electric-el@electric-eloquence.io>
Date:   Thu Jul 14 14:19:47 2022 -0400

    improved multiple content keys for front matter content

commit ac8c2297a433e231afd410b8c2fe2face13b5dea
Author: electric-el <electric-el@electric-eloquence.io>
Date:   Tue Jul 12 16:31:16 2022 -0400

    allowing multiple content keys for front matter content

commit 79760cfa67c8b774cf1a6aaefe34328adfec65f2
Author: electric-el <electric-el@electric-eloquence.io>
Date:   Mon Jul 11 15:37:53 2022 -0400

    sanitizing script tags posted to markdown editor

commit 5a6c6c49f1fd2c00112f6809cc0ef83f19286937
Author: electric-el <electric-el@electric-eloquence.io>
Date:   Sun Jul 10 17:51:09 2022 -0400

    minor css improvements to ui

commit 8aaa4ae4f2100a73d8ab4beab759e9fbe2718958
Author: electric-el <electric-el@electric-eloquence.io>
Date:   Thu Jul 7 18:36:39 2022 -0400

    added deprecation notices for listitems

commit 6a64334d6041a03412c3bfad26d54653f79e74ed
Author: electric-el <electric-el@electric-eloquence.io>
Date:   Thu Jul 7 18:14:36 2022 -0400

    fixes per linter

commit 2f1d2e3e6b277fb970bf919a5f4da2b3821f7b18
Author: electric-el <electric-el@electric-eloquence.io>
Date:   Thu Jul 7 18:12:34 2022 -0400

    restored deletion, undeprecated deprecation

commit 86e0a63c9703df7b4befae52d0fd751dd39a094b
Author: electric-el <electric-el@electric-eloquence.io>
Date:   Thu Jul 7 11:50:01 2022 -0400

    fixed bad copying in json compiler, passing tests

commit 0f7c46be64f2a8ca1793d68b2cf09c1ae6de2a34
Author: electric-el <electric-el@electric-eloquence.io>
Date:   Thu Jul 7 11:49:15 2022 -0400

    restored deletion, undeprecated deprecation

commit d0f792484b21498af2f7ea3bcb00d32d321eed45
Author: electric-el <electric-el@electric-eloquence.io>
Date:   Thu Jul 7 07:32:27 2022 -0400

    removed deprecated code

commit 13e4f6c53d66f7500f681f436e2fc6a97b7ebe55
Author: electric-el <electric-el@electric-eloquence.io>
Date:   Fri Jul 1 16:53:49 2022 -0400

    internationalizing requerio inspector text

commit bb4ee355c1124e2bc159e170907e2b5cd78a4569
Author: electric-el <electric-el@electric-eloquence.io>
Date:   Fri Jul 1 11:47:32 2022 -0400

    requerio inspector with text coloring

commit 89a8e4270fb318161d522f2257f532a58ea8bccb
Author: electric-el <electric-el@electric-eloquence.io>
Date:   Fri Jul 1 07:06:00 2022 -0400

    bumped he version

commit ccc79aa5e88afffd7e43b8d736436d22104bce0d
Author: electric-el <electric-el@electric-eloquence.io>
Date:   Tue Jun 28 05:54:16 2022 -0400

    updated styles for requerio inspector

commit fae52fc1f7802fb599cb6a6ea7a60da10c91def7
Author: electric-el <electric-el@electric-eloquence.io>
Date:   Sat Jun 25 06:14:57 2022 -0400

    requerio inspector html and css mostly set up

commit 06db1f437843476fb96d6e72f9b4a3b8bd5160b3
Author: electric-el <electric-el@electric-eloquence.io>
Date:   Fri Jun 24 16:16:50 2022 -0400

    requerio inspector html and css mostly set up

commit 3b477f0f9850386e26624dfbe18d2c92b564761c
Author: electric-el <electric-el@electric-eloquence.io>
Date:   Fri Jun 24 08:45:23 2022 -0400

    requerio inspector html and css mostly set up

commit 704517ab25bc432c8b93317967cce3f563b8a3c6
Author: electric-el <electric-el@electric-eloquence.io>
Date:   Fri Jun 24 05:38:45 2022 -0400

    style updates to ui

commit 7aede8d2d8adecf7ff2795040c22d175edfca545
Author: electric-el <electric-el@electric-eloquence.io>
Date:   Thu Jun 23 20:32:28 2022 -0400

    style updates to ui

commit a22ca40f3545445d7b8bee8cd9b88fea59854b13
Author: electric-el <electric-el@electric-eloquence.io>
Date:   Thu Jun 23 20:06:41 2022 -0400

    style updates to ui css

commit 2fe0d7c30aad39b7abfa19f54f16272eb1f18da2
Author: electric-el <electric-el@electric-eloquence.io>
Date:   Thu Jun 23 09:12:44 2022 -0400

    minor text update to code viewer html

commit 0bde3de9301da0d0a9d090bc576bb35573ebde70
Author: electric-el <electric-el@electric-eloquence.io>
Date:   Wed Jun 22 10:45:17 2022 -0400

    updated linting rules to accommodate es6 module

commit e68df0289609068271e0abb156f7ca9f482420b7
Author: electric-el <electric-el@electric-eloquence.io>
Date:   Wed Jun 22 10:38:47 2022 -0400

    some basic scaffolding for requerio inspector

+++ ui/core/lib/ui-compiler.js
@@ -244,10 +244,18 @@ module.exports = class {
     /* istanbul ignore if */
     if (fs.existsSync(`${this.config.paths.public.root}/${requerioDir}/${requerioFile}`)) {
       fs.ensureDirSync(`${this.config.paths.public.styleguide}/${requerioDir}`);
-      fs.copySync(
-        `${this.config.paths.public.root}/${requerioDir}/${requerioFile}`,
-        `${this.config.paths.public.styleguide}/${requerioDir}/${requerioFile}`
-      );
+
+      // Check that one Requerio file is not symbolically linked to another when developing and debugging.
+      if (
+        !fs.existsSync(`${this.config.paths.public.styleguide}/${requerioDir}/${requerioFile}`) ||
+        JSON.stringify(fs.statSync(`${this.config.paths.public.root}/${requerioDir}/${requerioFile}`)) !==
+        JSON.stringify(fs.statSync(`${this.config.paths.public.styleguide}/${requerioDir}/${requerioFile}`))
+      ) {
+        fs.copySync(
+          `${this.config.paths.public.root}/${requerioDir}/${requerioFile}`,
+          `${this.config.paths.public.styleguide}/${requerioDir}/${requerioFile}`
+        );
+      }
     }
 
     // Delete old compiled styles.
